Got rid of the std::variant based Message::Id, went with pure QVariant
It's less code to maintain to do that, plus it works with all compilers
(I had trouble with std::variant on non-Apple-clang).

On top of that the QVariant approach likely is slightly faster as no
conversion needs to be done from our Message::id -> QVariant when it
*is* a QVariant already.

We sacrifice a tiny bit of type safety.  We just have to be sure in the
code to never allow fractional doubles to go out as an id (this is
unlikely to happen anyway in practice).