C++ template magic - newTask factory function
Templates are fun.  Decided to reduce boilerplate by making task
creation inside the Controller class be done by a template member
function which acts like a task factory.  It surrounds the construction
of the task with some boilerplate (namely, inserting the task in the
tasks map, connecting an optional errored signal, and starting the
task).