Some tweaks for correctness and memory consumption + lru cache perf
- Added compactifyUtxoSet to periodically reduce all instances of dupe
QByteArrays to point to the same underlying shared object. This runs
every 33,000 blocks on synch.

- Merged the Hdr and UtxoSet save flags into a single flag, "Blocks"
which means they get saved together. This is crucial for correctness.
Before it could be the case that a different height would be saved for
headers vs utxo set because of the way it worked.

- made the two lru caches try and keep the same key/value pairs
(hopefully shared copies)

- made the LRU::Cache class not use an exception internally on tryGet()
-- this seems to have a positive impact on performance as this function
is called a lot.  Instead, it calls into an internal common function now
called get_nolock_nothrow, upon which everything else is based.

It appears these efforts have reduced memory consumption a bit.  Our
UTXOSet is still pretty heavy.  More work to be done...