Got rid of in-memory utxo set.. will rely on db
RocksDB is super fast. Keeping it in memory was a waste of memory.  This
is much faster and reduced memory consumption significantly.

TODO: Maybe do the same thing for headers as well!

Also: We need to save txNum to DB. We're not saving it now resulting in
correctness errors on subsequent reloads of same db.