Fix: Connection lost/reconnect detection logic in Controller.cpp
There was a race condition before if the controller thread was very busy
it would not have time to reconnect the "wait for auth" signal and as
such would end up in a state where it kept spamming "waiting for
bitcoind" to the log incessantly.

This has been handled by keeping the gotFirstGoodConnection signal
always connected and only reacting to it if the "lostConn" flag is
latched (latched by allConnectionsLost signal).

The gotFirstGoodConnection signal is only emitted when the first good
connection is made anyway so it's not spammed.  It was not ideal that we
didn't have it always connected.

This fix hopefully fixes the race condition we saw when synching mainnet
on my slow windows box.