Refactor: added the ShallowTmp and DeepCpy templates as QByteArray helpers
This will reduce typing and hopefully make it very explicit which type
of QByteArray is being created in the code as a defensive progarmming
measure to eliminate errors... because the QByteArray::FromRawData
static function doesn't capture its very unsafe nature in its name (and
the fact that it is a "contagious" property it brings to the
QByteArray).  The "ShallowTmp" naming captures this.