Using our new Merkle.cpp, fully implemented 2 of the transaction methods
blockchain.transaction.id_from_pos is now fully implemented, as is
blockchain.transaction.get_merkle.

They both use a cache to store the txHashes for a particular block, and
resort to reading from the disk file if not found in cache.

This cache stores the txHashes for 500 of the most recently queried
blocks.  In the worst pathological circumstances of 64k tx's per block,
the cache would use about 1.5GB of memory at worst.  In practice it will
use much less as blocks are empty.