Initial steps in making the mempool use less memory & CPU
- Made the hashXTxs mapped_type be a std::vector<TxRef> rather than a
std::set<TxRef>.  We manually maintain the uniqueness and sort invariant
as we process the mempool.  Vectors are much more compact memory-wise
and also have better insert characteristics, so we prefer them here.

Also, we never check for set membership, so the O log N lookup of the
set is not warranted.

We do however potentially build very large hashX -> TxRef arrays so a
std::vector is much better.

After this fix, we need to see what else in this data structure is
wasteful. I may get rid of the "txos" std::map<IONum, TXOInfo> in favor of a
simple and direct std::vector<TXOInfo>