Added: peering_enforce_unique_ip, plus nits and fixups
peering_enforce_unique_ip, if enabled, will have the PeerMgr keep track
of the IP address set and refuse to connect to peers from the same IP
even if they have multiple hostnames/ports.  This is the sybil attack
defense measure ElectronX uses.

Discovered std::as_const.  Will use it to block implicit sharing
detachment of some Qt containers.

Downgraded the "_sendError: Not connected!" message from an "Error" to a
Debug messaage.  This is really something we anticipated and doesn't
need to spam the log as an error.