Build: Make it possible to use a system supplied rocksdb
All the rocksdb headers have been moved to the staticlibs folder
to prevent include path conflicts. They have also been removed
from the project file as they are not necessary for building and
when using a system rocksdb the headers are different ones.

Storage.cpp has been changed to include rocksdb from the system
include path first.

In the project there are new feature flags, the first of them
being "staticlibs" and set by default. These can be overridden
from the qmake command line like this:

qmake -makefiles features= Fulcrum.pro

The project now only adds the staticlibs to the include and
library path when the feature is set.