Tweak to mempool: Improvement to detection of mempool drops and notification of drops
In the rare case of a mempool tx drop, we enqueue notifications for the
dropped tx's and reset the mempool (which also enqueues notifications
again for the entire mempool).

In practice this doesn't lead to very many status's being computed
because at any one time the probability that a mempool tx is one
belonging to a subscribed client is very small.

This commit now makes mempool drop detection very solid.

Previous to this commit it would be possible for a client to never be
motified of a drop in some cases.