Mempool synch optimization: Take a shared lock in the initial phase
We aren't modifying the mempool in the initial scan when we want to
detect drops, yet we were taking an exclusive lock.  This was
suboptimal.

Instead, we now scan the mempool to detect drops using a shared lock.
This way, clients can continue to read the mempool as we synch.

The initial scan was the slowest part of synching so this should help
performance under load.

Also we refactored the synchmempool task subs notification logic and moved it
into the task itself, simplifying the calling code in Controller
somewhat.

Tested, works great.