Fix: Printing to log of block totals was sometimes off
This is because we were tallying the totals form the 'success' callback
which wasn't always delivered in time before the statemachine for the dl
headers task was reset.

We instead tally the totals as we process the blocks.

Also updated the totals display to show number of addresses in processed
blocks as well.