Build: Add proper installation target
This will install documentation to $PREFIX/share/doc/$TARGET and
binaries to $PREFIX/bin. $PREFIX defaults to /usr/local and can be
overridden from the environment or qmake command line.

We need to override QMAKE_STRIP for the admin target because qmake
is calling strip on any file with execute bit enabled, which fails
for Python files.