Nit in AdminServer - bumped extant job limit to 100
Likely if the limit is reached, something bad is happening.  But this
limit was raised in case an admin uses the FulcrumAdmin script in
conjuntion with other automated scripts -- so that it has some
flexibility in backlogging many rapid resuests.  In practice I don't
envision the extant jobs going over 10.  But we went with 100 to be
safe.  Note that 100 is still a relatively conservative number.

The admin rpc is intentionally designed to be very resource-minimal and
not intended to serve many clients under load -- leaving room for the
public SPV server to take up all the resources. :)