PeerMgr: Added better heuristics for using IPv6-only hosts
A server admin, atroxes, pointed out that his server wasn't connecting
to IPv6-only hosts (such as electrumx-cash.1209k.com). It turns out the
heuristics we were using about deciding to try IPv6 AAAA records were
flawed.

We redid the logic to never give up and if the only thing resolved is
AAAA (IPv6), we do it anyway, even if we think we may not be IPv6
capable.

As the app runs it will learn if it really has IPv6 based on whether it
manages to connect to any ip6 peers.

The logic now should be what people expect and no peers should slip
through the cracks.