Synch. Progress Display: Replaced ins & outs/sec with txs/sec
For two reasons:

1. It takes fewer horizontal characters to say "txs/sec" than "ins &
outs/sec"

2. txs/sec is actually a more natural metric that people can think in
terms of.  The choice to display ins&outs/sec was purely when developing
to see what was the bottleneck in slow synch progress... (it turns out
it's a complicated set of factors mainly to do with huge address
histories being updated in some parts of the blockchain).

Anyway.. txs/sec is more useful to see as a stat when it synchs.