Bugfix: Ensure removal of zombie connections after idle timeout
It was possible in unlikely cases for client connections to "never go away"
and remain zombies forever.  This was due to the racey way that socket
events happen.  This has been fixed now -- sendError() always forces a
disconnect and cleanup of the client if the socket state appears
stale/disconnected.