Added the ability to change the server's max_buffer setting at runtime
Also in this commit: Set the default max_buffer back to 4000000.

Now we have the ability to set the max_buffer at runtime.

This is accomplished by two new FulcrumAdmin commands: `setmaxbuffer` for
setting and `maxbuffer` for reading.

The max_buffer will, in future commits, be more aggressively enforced
and clients exceeding it for a long enough period of time will be
kicked.