Implemented better kicking logic for clients filling the read buffer
This logic is normally never triggered by well-behaved EC clients, but I
have observed a new client with version string "3.7.11" abusing the
server and filling the read buffers.

This is not critical -- but is a waste of resources (and potentially
bandwidth) to serve such clients.

So we kick them if they exceed the read buffer size (default 4MiB) for
longer than 5 seconds.

This, combined with the new throttling logic, should make the server
more resilient to misbehaving clients.