Implemented blockchain.address.* methods
This involved some slight refactoring to refactor out the common code
used by both blockchain.scripthash.* and blockchain.address.*, but
overall it's a very clean implementation.

Note that blockchain.address.subscribe will always notify on the exact
string (trimmed) that the client supplied.  This will need to be
documented in the protocol spec.