Static build: added build target for older Linux (linux_ub16)
This allows us to entirely build releases using Docker.  The linux_ub16
target is for older systems with an older libc (Ubuntu16 LTS or
thereabouts), which can't run the newer static build (built on top of
Ubuntu 18 LTS but it requires a newer libc than Ubuntu16 has).

Also in this commit we simplified the Linux Dockerfile to not install
dependencies we don't need, and updated the README to reflect the new
platform targets for build.sh.

The platform targets for build.sh now are: linux, windows, and
linux_ub16 (also short versions work: lin, win, oldlin).