AbstractConnection: Improved "writeBackLog" and MAX_BUFFER enforcement
Our previous "writeBackLog" and/or MAX_BUFFER logic for writes was flawed
(but not critically so).

We were creating a QByteArray buffer previously that was never used,
since careful reading of the QTcpSocket code shows that when it does
write, it uses an unlimited write buffer and always succeeds fully.

So -- in order to detect how big the QTcpSocket write backlog really is,
we instead rely on the bytesWritten() signal to tell us when a write
actually completes and we decrement the writeBackLog counter at that
point.

After this change, AbstractConnection::do_write now has logic to detect
the situation where the write backlog is actually beyond MAX_BUFFER, and
in that case, we error-out the client and kill the connection.