SSL Server: Improved handshake mechanism
We added a 10 second timeout to the TLS handshake phase upon receiving a
connection from an SSL client. This prevents situations where bad clients
can keep sockets open for too long during the TLS handshake phase for the
QSslSocket in ServerSSL (in the incomingConnection slot).

Now, after a 10 second grace period, if the handshake hasn't yet completed,
we kill the connection. Also, if SSL errors are encountered during the
handshake, we abort the connection as well.

We also check for race-conditioney type situations such as the
connection entering the disconnected state from under us (this shouldn't
happen anyway due to the way connection states are managed from within
the event loop, but we do this anyway for correctness).