Added WebSockets classes and low-level code
Implemented RFC 6455 from: https://tools.ietf.org/html/rfc6455

Our implementation lives in WebSockets.h and WebSockets.cpp. It provides
low-level code for serializing/deserializing data to/from the wire as
well as a higher level "Wrapper" class that derives from QTcpSocket and
is used to wrap an existing QTcpSocket as a facade.

The WebSockets::Wrapper is intented to be used asynchronously via the
readyRead() or messagesReady() signal.

Writing to the WebSocket can be accomplished via the write() method (as
with a regular QTcpSocket), in which case the entire QByteArray buffer
is sent down the wire as a single message (possibly fragmented) of
either Text or Binary type (default Text).

Additionally, custom methods sendText() and sendBinary() are provided.

PING/PONG support is provided via an optional automatic mechanism (on by
default).  CLOSE is handled automatically as well.