Slight optimmization to how we use std::optional, app-wide
It turns out operator*() and operator->() on std::optional is faster
than .value() if you know for a fact the optional has_value() (these
operators don't check if has_value and/or don't throw -- they do an unsafe
get() for you).

So, app-wide I replaced usages of opt.value() with *opt or
opt->member wherever I could, for the slight performance benefit that
gives us.