Fixed CA signed certs to actually work
It turns out we needed to load the certificate chain (fullchain.pem). CA
signed certs should now work as advertised.  Also updated the help to
reflect this.

The app detects if the cert is self-signed or not, and if it is not (CA
signed), it insists on being given the fullchain.pem file.

Also -- added a custom qt message handler to suppress some annoying qt
log messages.

Updated man page as well.