Hashing: Fixed-up/cleaned-up the hashers used for maps, etc.
- Our hashers weren't very good. They were trying to be really fast at
the expense of risking collisions.  We instead always hash all ints, etc
using the C++ hasher (murmur2 or city hash) or qHash, depending on
context.

- Our hashers were not terribly good for 32-bit platforms

- We made assumptions about being able to cast pointers to data to
uint64 which is a bad assumption to make. (On some platforms such as
older ARM CPUs this can lead to alignment issues). This has been addressed.